Use env vars to pass host and not exclusion filters
Exclusion filters should not be used to pass
the host name. They are an rspec feature that
is not intended to pass arbitrary information.

In my use case, where I need to specify ip
addresses as the address.

This patch updates the Rakefile and spec_helper
to pass hostnames as the TARGET_HOST env var (
which matches the multi-host example from
http://serverspec.org/advanced_tips.html